+++ src/main/java/com/example/demo/services/CalculatorService.java
@@ -1,12 +1,42 @@
 package com.example.demo.services;
 
+import org.springframework.stereotype.Service;
 import org.springframework.ui.Model;
 
-public interface CalculatorService {
-    String index();
-    String sumar(int num1, int num2, Model model);
-    String restar(int num1, int num2, Model model);
-    String multiplicar(int num1, int num2, Model model);
-    String dividir(int num1, int num2, Model model);
+@Service
+public class CalculatorService {
+
+    public String index() {
+        return "index";
+    }
+
+    public String sumar(int num1, int num2, Model model) {
+        double resultado = num1 + num2;
+        model.addAttribute("resultado", resultado);
+        return "index";
+    }
+
+    public String restar(int num1, int num2, Model model) {
+        double resultado = num1 - num2;
+        model.addAttribute("resultado", resultado);
+        return "index";
+    }
+
+    public String multiplicar(int num1, int num2, Model model) {
+        double resultado = num1 * num2;
+        model.addAttribute("resultado", resultado);
+        return "index";
+    }
+
+    public String dividir(int num1, int num2, Model model) {
+
+        if (num2 != 0) {
+            double resultado = (double) num1 / num2;
+            model.addAttribute("resultado", resultado);
+        } else {
+            model.addAttribute("error", "No se puede dividir por cero");
+        }
+        return "index";
+    }
 
 }
